Split out DataSerializableInternal interface.
