+++ cppcache/include/geode/internal/DataSerializableInternal.hpp
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#ifndef GEODE_DATASERIALIZABLEINTERNAL_H_
+#define GEODE_DATASERIALIZABLEINTERNAL_H_
+
+#include "../Serializable.hpp"
+
+namespace apache {
+namespace geode {
+namespace client {
+
+class DataOutput;
+class DataInput;
+
+namespace internal {
+
+class APACHE_GEODE_EXPORT DataSerializableInternal
+    : public virtual Serializable {
+ public:
+  ~DataSerializableInternal() override = default;
+  virtual void toData(DataOutput &dataOutput) const = 0;
+  virtual void fromData(DataInput &dataInput) = 0;
+  virtual int8_t getInternalId() const = 0;
+};
+
+}  // namespace internal
+}  // namespace client
+}  // namespace geode
+}  // namespace apache
+
+#endif  // GEODE_DATASERIALIZABLEINTERNAL_H_
