+++ cppcache/src/PdxFieldType.hpp
@@ -23,6 +23,7 @@
 #include <string>
 
 #include <geode/internal/geode_globals.hpp>
+#include <geode/internal/DataSerializableInternal.hpp>
 #include <geode/Serializable.hpp>
 #include <geode/CacheableString.hpp>
 #include <geode/DataInput.hpp>
@@ -33,7 +34,8 @@ namespace apache {
 namespace geode {
 namespace client {
 
-class APACHE_GEODE_EXPORT PdxFieldType : public DataSerializableInternal {
+class APACHE_GEODE_EXPORT PdxFieldType
+    : public internal::DataSerializableInternal {
  private:
   std::string m_fieldName;
   std::string m_className;
@@ -51,8 +53,9 @@ class APACHE_GEODE_EXPORT PdxFieldType : public DataSerializableInternal {
   int32_t getFixedTypeSize() const;
 
  public:
-  PdxFieldType(std::string fieldName, std::string className, PdxFieldTypes typeId,
-               int32_t sequenceId, bool isVariableLengthType, int32_t fixedSize,
+  PdxFieldType(std::string fieldName, std::string className,
+               PdxFieldTypes typeId, int32_t sequenceId,
+               bool isVariableLengthType, int32_t fixedSize,
                int32_t varLenFieldIdx);
 
   PdxFieldType();
